requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.2
WTForms==3.1.2
Flask-WTF==1.2.1
razorpay
google-auth
python-dotenv
requests
